[refactor/TK-215]: 토큰 재발급 로직을 AuthService로 이동
