--- src/main/java/withbeetravel/service/auth/RefreshTokenService.java
@@ -1,12 +1,0 @@
-package withbeetravel.service.auth;
-
-import withbeetravel.dto.request.auth.RefreshTokenDto;
-import withbeetravel.dto.response.auth.ExpirationDto;
-import withbeetravel.dto.response.auth.SignInResponseDto;
-
-public interface RefreshTokenService {
-
-    SignInResponseDto refreshToken(final String refreshToken);
-
-    ExpirationDto checkExpirationTime(RefreshTokenDto refreshTokenDto);
-}
